Add findActionButton helped globally for tests
